adds currency iso code as separate value object
