+++ src/Adapter/Currency/CommandHandler/CreateCurrencyHandler.php
@@ -50,11 +50,10 @@ final class CreateCurrencyHandler extends AbstractObjectModelLegacyHandler imple
     public function handle(CreateCurrencyCommand $command)
     {
         //todo: check if currency does not exist
-        //todo: try to enter 0.000000001
         try {
             $entity = new Currency();
 
-            $entity->iso_code = $command->getIsoCode();
+            $entity->iso_code = $command->getIsoCode()->getValue();
             $entity->active = $command->isEnabled();
             $entity->conversion_rate = $command->getExchangeRate();
