+++ src/Core/Form/IdentifiableObject/DataHandler/CurrencyFormDataHandler.php
@@ -30,6 +30,7 @@ use PrestaShop\PrestaShop\Core\CommandBus\CommandBusInterface;
 use PrestaShop\PrestaShop\Core\Domain\Currency\Command\CreateCurrencyCommand;
 use PrestaShop\PrestaShop\Core\Domain\Currency\Command\UpdateCurrencyCommand;
 use PrestaShop\PrestaShop\Core\Domain\Currency\ValueObject\CurrencyId;
+use PrestaShop\PrestaShop\Core\Domain\Currency\ValueObject\IsoCode;
 
 /**
  * Class CurrencyFormDataHandler
@@ -54,7 +55,12 @@ final class CurrencyFormDataHandler implements FormDataHandlerInterface
      */
     public function create(array $data)
     {
-        $command = new CreateCurrencyCommand($data['iso_code'], $data['exchange_rate'], $data['active']);
+        $command = new CreateCurrencyCommand(
+            new IsoCode($data['iso_code']),
+            $data['exchange_rate'],
+            $data['active']
+        );
+
         $command->setShopIds($data['shop_association']);
 
         /** @var CurrencyId $currencyId */
